db utility commands and restructure
